Move snes specific code to a sub directory
