+++ WebRandomizer/ClientApp/src/snes/item_lookup.js
